Adjust exports to simplify testing setup
